proximity function adjustments

+++ pkg/kademlia/kademlia.go
@@ -169,7 +169,7 @@ func (k *Kad) manage() {
 
 			if err != nil {
 				if errors.Is(err, errMissingAddressBookEntry) {
-					po := uint8(swarm.Proximity(k.base.Bytes(), peerToRemove.Bytes()))
+					po := swarm.Proximity(k.base.Bytes(), peerToRemove.Bytes())
 					k.knownPeers.Remove(peerToRemove, po)
 				} else {
 					k.logger.Errorf("kademlia manage loop iterator: %v", err)
@@ -292,7 +292,7 @@ func (k *Kad) AddPeer(ctx context.Context, addr swarm.Address) error {
 	}
 
 	po := swarm.Proximity(k.base.Bytes(), addr.Bytes())
-	k.knownPeers.Add(addr, uint8(po))
+	k.knownPeers.Add(addr, po)
 
 	select {
 	case k.manageC <- struct{}{}:
@@ -304,7 +304,7 @@ func (k *Kad) AddPeer(ctx context.Context, addr swarm.Address) error {
 
 // Connected is called when a peer has dialed in.
 func (k *Kad) Connected(ctx context.Context, addr swarm.Address) error {
-	po := uint8(swarm.Proximity(k.base.Bytes(), addr.Bytes()))
+	po := swarm.Proximity(k.base.Bytes(), addr.Bytes())
 	k.knownPeers.Add(addr, po)
 	k.connectedPeers.Add(addr, po)
 
@@ -325,7 +325,7 @@ func (k *Kad) Connected(ctx context.Context, addr swarm.Address) error {
 
 // Disconnected is called when peer disconnects.
 func (k *Kad) Disconnected(addr swarm.Address) {
-	po := uint8(swarm.Proximity(k.base.Bytes(), addr.Bytes()))
+	po := swarm.Proximity(k.base.Bytes(), addr.Bytes())
 	k.connectedPeers.Remove(addr, po)
 
 	k.waitNextMu.Lock()
